AP_Networking: fixed duplicate parameters in wiki
defining two parameter tables in one file causes duplicates
